saulgpt/SaulGPT/backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "saulgpt-backend"
version = "0.1.0"
description = "FastAPI backend for SaulGPT - RAG-powered legal assistant"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-ollama>=0.2.0",
    "pgvector>=0.3.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.0.0",
    "sentence-transformers>=3.0.0",
    "langchain-text-splitters>=0.3.0",
    "pypdf>=5.0.0",
    "python-docx>=1.1.0",
    "reportlab>=4.0.0",
    "python-multipart>=0.0.9",
    "ai-datastream>=0.1.0",
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "httpx>=0.27.0"]

[tool.setuptools]
packages = ["app"]
